+++ src/databricks/labs/ucx/contexts/application.py
@@ -31,6 +31,11 @@ from databricks.labs.ucx.hive_metastore.table_move import TableMove
 from databricks.labs.ucx.hive_metastore.udfs import UdfsCrawler
 from databricks.labs.ucx.hive_metastore.verification import VerifyHasMetastore
 from databricks.labs.ucx.installer.workflows import DeployedWorkflows
+from databricks.labs.ucx.source_code.notebooks.loaders import NotebookResolver, NotebookLoader, WorkspaceNotebookLoader
+from databricks.labs.ucx.source_code.files import FileLoader, LocalFileResolver, SysPathProvider
+from databricks.labs.ucx.source_code.graph import DependencyResolver, DependencyGraphBuilder
+from databricks.labs.ucx.source_code.whitelist import WhitelistResolver, Whitelist
+from databricks.labs.ucx.source_code.site_packages import SitePackagesResolver, SitePackages
 from databricks.labs.ucx.source_code.languages import Languages
 from databricks.labs.ucx.workspace_access import generic, redash
 from databricks.labs.ucx.workspace_access.groups import GroupManager
@@ -334,3 +339,56 @@ class GlobalContext(abc.ABC):
     @cached_property
     def verify_has_metastore(self):
         return VerifyHasMetastore(self.workspace_client)
+
+    @cached_property
+    def notebook_loader(self) -> NotebookLoader:
+        return WorkspaceNotebookLoader(self.workspace_client)
+
+    @cached_property
+    def notebook_resolver(self):
+        return NotebookResolver(self.notebook_loader)
+
+    @cached_property
+    def site_packages(self):
+        # TODO: actually load the site packages
+        return SitePackages([])
+
+    @cached_property
+    def syspath_provider(self):
+        return SysPathProvider.from_sys_path()
+
+    @cached_property
+    def file_loader(self):
+        return FileLoader(self.syspath_provider)
+
+    @cached_property
+    def site_packages_resolver(self):
+        return SitePackagesResolver(self.site_packages, self.file_loader, self.syspath_provider)
+
+    @cached_property
+    def whitelist(self):
+        # TODO: fill in the whitelist
+        return Whitelist()
+
+    @cached_property
+    def whitelist_resolver(self):
+        return WhitelistResolver(self.whitelist)
+
+    @cached_property
+    def file_resolver(self):
+        return LocalFileResolver(self.file_loader)
+
+    @cached_property
+    def dependency_resolver(self):
+        return DependencyResolver(
+            [
+                self.notebook_resolver,
+                self.site_packages_resolver,
+                self.whitelist_resolver,
+                self.file_resolver,
+            ]
+        )
+
+    @cached_property
+    def dependency_graph_builder(self):
+        return DependencyGraphBuilder(self.dependency_resolver)
