+++ src/databricks/labs/ucx/source_code/base.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from abc import abstractmethod
 from collections.abc import Iterable
 from dataclasses import dataclass
