+++ tests/unit/__init__.py
@@ -18,9 +18,10 @@ from databricks.sdk.service.sql import EndpointConfPair
 from databricks.sdk.service.workspace import ExportResponse, GetSecretResponse, Language
 
 from databricks.labs.ucx.hive_metastore.mapping import TableMapping, TableToMigrate
-from databricks.labs.ucx.source_code.dependency_loaders import SourceContainer
-from databricks.labs.ucx.source_code.files import LocalFile
-from databricks.labs.ucx.source_code.notebook import Notebook, NOTEBOOK_HEADER
+from databricks.labs.ucx.source_code.graph import SourceContainer
+from databricks.labs.ucx.source_code.files import LocalFile, FileLoader, SysPathProvider
+from databricks.labs.ucx.source_code.notebooks.sources import Notebook
+from databricks.labs.ucx.source_code.notebooks.cells import NOTEBOOK_HEADER
 from databricks.labs.ucx.source_code.whitelist import Whitelist
 
 logging.getLogger("tests").setLevel("DEBUG")
@@ -132,6 +133,10 @@ def _load_sources(cls: type, *filenames: str):
     return [installation.load(cls, filename=_) for _ in filenames]
 
 
+def _samples_path(cls: type) -> str:
+    return (__dir / _FOLDERS[cls]).as_posix()
+
+
 def _cluster_policy(policy_id: str):
     fixture = _load_fixture(f"{_FOLDERS[Policy]}/{policy_id}.json")
     definition = json.dumps(fixture["definition"])
@@ -195,7 +200,7 @@ def _load_dependency_side_effect(sources: dict[str, str], visited: dict[str, boo
 
 def _is_notebook_side_effect(sources: dict[str, str], *args):
     dependency = args[0]
-    filename = str(dependency.path)
+    filename = dependency.path.as_posix()
     if filename.startswith('./'):
         filename = filename[2:]
     source = sources.get(filename, None)
@@ -207,6 +212,59 @@ def _is_notebook_side_effect(sources: dict[str, str], *args):
         source = sources.get(filename)
     assert source is not None
     return NOTEBOOK_HEADER in source
+
+
+def _full_path_side_effect(sources: dict[str, str], *args):
+    path = args[0]
+    filename = path.as_posix()
+    if filename.startswith('./'):
+        filename = filename[2:]
+    if filename in sources:
+        return pathlib.Path(filename)
+    if filename.find(".py") < 0:
+        filename = filename + ".py"
+    if filename.find(".txt") < 0:
+        filename = filename + ".txt"
+    if filename in sources:
+        return pathlib.Path(filename)
+    return None
+
+
+def _is_file_side_effect(sources: dict[str, str], *args):
+    return _full_path_side_effect(sources, *args) is not None
+
+
+def _local_loader_with_side_effects(cls: type, sources: dict[str, str], visited: dict[str, bool]):
+    file_loader = create_autospec(cls)
+    file_loader.exists.side_effect = lambda *args, **kwargs: _is_file_side_effect(sources, *args)
+    file_loader.is_notebook.return_value = False
+    file_loader.full_path.side_effect = lambda *args: _full_path_side_effect(sources, *args)
+    file_loader.load_dependency.side_effect = lambda *args, **kwargs: _load_dependency_side_effect(
+        sources, visited, *args
+    )
+    return file_loader
+
+
+class TestFileLoader(FileLoader):
+    __test__ = False
+
+    def __init__(self, syspath_provider: SysPathProvider, sources: dict[str, str]):
+        super().__init__(syspath_provider)
+        self._sources = sources
+
+    def exists(self, path: pathlib.Path):
+        if super().exists(path):
+            return True
+        filename = path.as_posix()
+        if filename.startswith('./'):
+            filename = filename[2:]
+        if filename in self._sources:
+            return True
+        if filename.find(".py") < 0:
+            filename = filename + ".py"
+        if filename.find(".txt") < 0:
+            filename = filename + ".txt"
+        return filename in self._sources
 
 
 def workspace_client_mock(
