+++ tests/unit/source_code/test_dependencies.py
@@ -1,29 +1,32 @@
-import os.path
 from pathlib import Path
 from unittest.mock import create_autospec
 
 from databricks.sdk import WorkspaceClient
 from databricks.sdk.service.workspace import ObjectInfo, Language, ObjectType
 
-from databricks.labs.ucx.source_code.dependencies import (
-    DependencyGraphBuilder,
-    DependencyProblem,
-)
-from databricks.labs.ucx.source_code.dependency_loaders import (
+from databricks.labs.ucx.source_code.graph import (
     SourceContainer,
-    LocalFileLoader,
-    LocalNotebookLoader,
-    WorkspaceNotebookLoader,
+    DependencyResolver,
+    DependencyProblem,
+    DependencyGraphBuilder,
 )
-from databricks.labs.ucx.source_code.dependency_resolvers import DependencyResolver
-from databricks.labs.ucx.source_code.site_packages import SitePackages
-from databricks.labs.ucx.source_code.whitelist import Whitelist
+from databricks.labs.ucx.source_code.notebooks.loaders import (
+    NotebookResolver,
+    WorkspaceNotebookLoader,
+    LocalNotebookLoader,
+)
+from databricks.labs.ucx.source_code.files import FileLoader, LocalFileResolver, SysPathProvider
+from databricks.labs.ucx.source_code.whitelist import WhitelistResolver, Whitelist
+from databricks.labs.ucx.source_code.site_packages import SitePackagesResolver, SitePackages
 from tests.unit import (
     _load_sources,
     _download_side_effect,
     whitelist_mock,
     _load_dependency_side_effect,
     locate_site_packages,
+    TestFileLoader,
+    _samples_path,
+    _local_loader_with_side_effects,
 )
 
 
@@ -39,11 +42,13 @@ def test_dependency_graph_builder_visits_workspace_notebook_dependencies():
     ws = create_autospec(WorkspaceClient)
     ws.workspace.download.side_effect = lambda *args, **kwargs: _download_side_effect(sources, visited, *args, **kwargs)
     ws.workspace.get_status.side_effect = get_status_side_effect
-    whi = whitelist_mock()
-    file_loader = create_autospec(LocalFileLoader)
-    file_loader.is_notebook.return_value = False
-    site_packages = SitePackages.parse(locate_site_packages())
-    builder = DependencyGraphBuilder(DependencyResolver(whi, site_packages, file_loader, WorkspaceNotebookLoader(ws)))
+    workspace_notebook_loader = WorkspaceNotebookLoader(ws)
+    dependency_resolver = DependencyResolver(
+        [
+            NotebookResolver(workspace_notebook_loader),
+        ]
+    )
+    builder = DependencyGraphBuilder(dependency_resolver)
     builder.build_notebook_dependency_graph(Path("root3.run.py.txt"))
     assert len(visited) == len(paths)
 
@@ -53,20 +58,29 @@ def test_dependency_graph_builder_visits_local_notebook_dependencies():
     sources: dict[str, str] = dict(zip(paths, _load_sources(SourceContainer, *paths)))
     visited: dict[str, bool] = {}
     whi = whitelist_mock()
-    file_loader = create_autospec(LocalFileLoader)
+    file_loader = create_autospec(FileLoader)
     file_loader.load_dependency.side_effect = lambda *args, **kwargs: _load_dependency_side_effect(
         sources, visited, *args
     )
     file_loader.is_notebook.return_value = True
-    file_loader.is_file.return_value = True
+    file_loader.exists.return_value = True
     notebook_loader = create_autospec(LocalNotebookLoader)
     notebook_loader.load_dependency.side_effect = lambda *args, **kwargs: _load_dependency_side_effect(
         sources, visited, *args
     )
     notebook_loader.is_notebook.return_value = True
-    notebook_loader.is_file.return_value = True
+    notebook_loader.exists.return_value = True
     site_packages = SitePackages.parse(locate_site_packages())
-    builder = DependencyGraphBuilder(DependencyResolver(whi, site_packages, file_loader, notebook_loader))
+    provider = SysPathProvider.from_pathlike_string("")
+    dependency_resolver = DependencyResolver(
+        [
+            NotebookResolver(notebook_loader),
+            SitePackagesResolver(site_packages, file_loader, provider),
+            WhitelistResolver(whi),
+            LocalFileResolver(file_loader),
+        ]
+    )
+    builder = DependencyGraphBuilder(dependency_resolver)
     builder.build_notebook_dependency_graph(Path("root4.py.txt"))
     assert len(visited) == len(paths)
 
@@ -75,27 +89,21 @@ def test_dependency_graph_builder_visits_workspace_file_dependencies():
     paths = ["root8.py.txt", "leaf1.py.txt", "leaf2.py.txt"]
     sources: dict[str, str] = dict(zip(paths, _load_sources(SourceContainer, *paths)))
     visited: dict[str, bool] = {}
-
-    def get_status_side_effect(*args):
-        path = args[0]
-        return (
-            ObjectInfo(path=path, object_type=ObjectType.FILE)
-            if path.startswith("leaf")
-            else ObjectInfo(path=path, language=Language.PYTHON, object_type=ObjectType.NOTEBOOK)
-        )
-
-    ws = create_autospec(WorkspaceClient)
-    ws.workspace.download.side_effect = lambda *args, **kwargs: _download_side_effect(sources, visited, *args, **kwargs)
-    ws.workspace.get_status.side_effect = get_status_side_effect
+    file_loader = _local_loader_with_side_effects(FileLoader, sources, visited)
     whi = whitelist_mock()
-    file_loader = create_autospec(LocalFileLoader)
-    file_loader.is_notebook.return_value = False
-    file_loader.load_dependency.side_effect = lambda *args, **kwargs: _load_dependency_side_effect(
-        sources, visited, *args
-    )
     site_packages = SitePackages.parse(locate_site_packages())
-    builder = DependencyGraphBuilder(DependencyResolver(whi, site_packages, file_loader, WorkspaceNotebookLoader(ws)))
-    builder.build_local_file_dependency_graph(Path("root8.py.txt"))
+    provider = SysPathProvider.from_pathlike_string("")
+    notebook_loader = LocalNotebookLoader(provider)
+    dependency_resolver = DependencyResolver(
+        [
+            NotebookResolver(notebook_loader),
+            SitePackagesResolver(site_packages, file_loader, provider),
+            WhitelistResolver(whi),
+            LocalFileResolver(file_loader),
+        ]
+    )
+    builder = DependencyGraphBuilder(dependency_resolver)
+    builder.build_local_file_dependency_graph(Path(paths[0]))
     assert len(visited) == len(paths)
 
 
@@ -110,17 +118,28 @@ def test_dependency_graph_builder_raises_problem_with_unfound_workspace_notebook
         return ObjectInfo(object_type=ObjectType.NOTEBOOK, language=Language.PYTHON, path=path)
 
     ws = create_autospec(WorkspaceClient)
+    ws.workspace.is_notebook.return_value = True
     ws.workspace.download.side_effect = lambda *args, **kwargs: _download_side_effect(sources, {}, *args, **kwargs)
     ws.workspace.get_status.side_effect = get_status_side_effect
     whi = whitelist_mock()
-    file_loader = create_autospec(LocalFileLoader)
+    file_loader = create_autospec(FileLoader)
     file_loader.is_notebook.return_value = False
     site_packages = SitePackages.parse(locate_site_packages())
-    builder = DependencyGraphBuilder(DependencyResolver(whi, site_packages, file_loader, WorkspaceNotebookLoader(ws)))
+    provider = SysPathProvider.from_pathlike_string("")
+    notebook_loader = WorkspaceNotebookLoader(ws)
+    dependency_resolver = DependencyResolver(
+        [
+            NotebookResolver(notebook_loader),
+            SitePackagesResolver(site_packages, file_loader, provider),
+            WhitelistResolver(whi),
+            LocalFileResolver(file_loader),
+        ]
+    )
+    builder = DependencyGraphBuilder(dependency_resolver)
     builder.build_notebook_dependency_graph(Path("root1.run.py.txt"))
-    assert builder.problems == [
+    assert list(builder.problems) == [
         DependencyProblem(
-            'dependency-check', 'Notebook not found: leaf2.py.txt', Path("root1.run.py.txt"), 19, 0, 19, 21
+            'notebook-not-found', 'Notebook not found: leaf2.py.txt', Path("root1.run.py.txt"), 19, 0, 19, 21
         )
     ]
 
@@ -134,19 +153,28 @@ def test_dependency_graph_builder_raises_problem_with_unfound_local_notebook_dep
         path = args[0]
         return path.as_posix() in paths
 
-    file_loader = create_autospec(LocalFileLoader)
-    file_loader.is_file.side_effect = is_file_side_effect
+    file_loader = create_autospec(FileLoader)
+    file_loader.exists.side_effect = is_file_side_effect
     file_loader.is_notebook.side_effect = is_file_side_effect
     file_loader.load_dependency.side_effect = lambda *args: _load_dependency_side_effect(sources, {}, *args)
     notebook_loader = create_autospec(LocalNotebookLoader)
-    notebook_loader.is_file.side_effect = is_file_side_effect
+    notebook_loader.exists.side_effect = is_file_side_effect
     notebook_loader.is_notebook.side_effect = is_file_side_effect
     notebook_loader.load_dependency.side_effect = lambda *args: _load_dependency_side_effect(sources, {}, *args)
     site_packages = SitePackages.parse(locate_site_packages())
-    builder = DependencyGraphBuilder(DependencyResolver(whi, site_packages, file_loader, notebook_loader))
+    provider = SysPathProvider.from_pathlike_string("")
+    dependency_resolver = DependencyResolver(
+        [
+            NotebookResolver(notebook_loader),
+            SitePackagesResolver(site_packages, file_loader, provider),
+            WhitelistResolver(whi),
+            LocalFileResolver(file_loader),
+        ]
+    )
+    builder = DependencyGraphBuilder(dependency_resolver)
     builder.build_notebook_dependency_graph(Path(paths[0]))
-    assert builder.problems == [
-        DependencyProblem('dependency-check', 'Notebook not found: leaf3.py.txt', Path(paths[0]), 1, 0, 1, 38)
+    assert list(builder.problems) == [
+        DependencyProblem('notebook-not-found', 'Notebook not found: leaf3.py.txt', Path(paths[0]), 1, 0, 1, 38)
     ]
 
 
@@ -159,20 +187,29 @@ def test_dependency_graph_builder_raises_problem_with_non_constant_local_noteboo
         path = args[0]
         return path.as_posix() in paths
 
-    file_loader = create_autospec(LocalFileLoader)
-    file_loader.is_file.side_effect = is_file_side_effect
+    file_loader = create_autospec(FileLoader)
+    file_loader.exists.side_effect = is_file_side_effect
     file_loader.is_notebook.side_effect = is_file_side_effect
     file_loader.load_dependency.side_effect = lambda *args: _load_dependency_side_effect(sources, {}, *args)
     notebook_loader = create_autospec(LocalNotebookLoader)
-    notebook_loader.is_file.side_effect = is_file_side_effect
+    notebook_loader.exists.side_effect = is_file_side_effect
     notebook_loader.is_notebook.side_effect = is_file_side_effect
     notebook_loader.load_dependency.side_effect = lambda *args: _load_dependency_side_effect(sources, {}, *args)
     site_packages = SitePackages.parse(locate_site_packages())
-    builder = DependencyGraphBuilder(DependencyResolver(whi, site_packages, file_loader, notebook_loader))
+    provider = SysPathProvider.from_pathlike_string("")
+    dependency_resolver = DependencyResolver(
+        [
+            NotebookResolver(notebook_loader),
+            SitePackagesResolver(site_packages, file_loader, provider),
+            WhitelistResolver(whi),
+            LocalFileResolver(file_loader),
+        ]
+    )
+    builder = DependencyGraphBuilder(dependency_resolver)
     builder.build_notebook_dependency_graph(Path(paths[0]))
-    assert builder.problems == [
+    assert list(builder.problems) == [
         DependencyProblem(
-            'dependency-check', "Can't check dependency not provided as a constant", Path(paths[0]), 2, 0, 2, 35
+            'dependency-not-constant', "Can't check dependency not provided as a constant", Path(paths[0]), 2, 0, 2, 35
         )
     ]
 
@@ -186,14 +223,46 @@ def test_dependency_graph_builder_raises_problem_with_invalid_run_cell():
         object_type=ObjectType.NOTEBOOK, language=Language.PYTHON, path=paths[0]
     )
     whi = whitelist_mock()
-    file_loader = create_autospec(LocalFileLoader)
+    file_loader = create_autospec(FileLoader)
     file_loader.is_notebook.return_value = False
     site_packages = SitePackages.parse(locate_site_packages())
-    builder = DependencyGraphBuilder(DependencyResolver(whi, site_packages, file_loader, WorkspaceNotebookLoader(ws)))
+    provider = SysPathProvider.from_pathlike_string("")
+    notebook_loader = WorkspaceNotebookLoader(ws)
+    dependency_resolver = DependencyResolver(
+        [
+            NotebookResolver(notebook_loader),
+            SitePackagesResolver(site_packages, file_loader, provider),
+            WhitelistResolver(whi),
+            LocalFileResolver(file_loader),
+        ]
+    )
+    builder = DependencyGraphBuilder(dependency_resolver)
     builder.build_notebook_dependency_graph(Path(paths[0]))
-    assert builder.problems == [
-        DependencyProblem('dependency-check', 'Missing notebook path in %run command', Path("leaf6.py.txt"), 5, 0, 5, 4)
+    assert list(builder.problems) == [
+        DependencyProblem('invalid-run-cell', 'Missing notebook path in %run command', Path("leaf6.py.txt"), 5, 0, 5, 4)
     ]
+
+
+def test_dependency_graph_builder_visits_recursive_file_dependencies():
+    paths = ["root6.py.txt", "root5.py.txt", "leaf4.py.txt"]
+    sources: dict[str, str] = dict(zip(paths, _load_sources(SourceContainer, *paths)))
+    visited: dict[str, bool] = {}
+    whi = whitelist_mock()
+    file_loader = _local_loader_with_side_effects(FileLoader, sources, visited)
+    site_packages = SitePackages.parse(locate_site_packages())
+    provider = SysPathProvider.from_pathlike_string("")
+    notebook_loader = LocalNotebookLoader(provider)
+    dependency_resolver = DependencyResolver(
+        [
+            NotebookResolver(notebook_loader),
+            SitePackagesResolver(site_packages, file_loader, provider),
+            WhitelistResolver(whi),
+            LocalFileResolver(file_loader),
+        ]
+    )
+    builder = DependencyGraphBuilder(dependency_resolver)
+    builder.build_local_file_dependency_graph(Path("root6.py.txt"))
+    assert len(visited) == len(paths)
 
 
 def test_dependency_graph_builder_raises_problem_with_unresolved_import():
@@ -201,23 +270,23 @@ def test_dependency_graph_builder_raises_problem_with_unresolved_import():
     sources: dict[str, str] = dict(zip(paths, _load_sources(SourceContainer, *paths)))
     visited: dict[str, bool] = {}
     whi = whitelist_mock()
-
-    def is_file_side_effect(*args):
-        path = args[0]
-        return str(path) in paths
-
-    file_loader = create_autospec(LocalFileLoader)
-    file_loader.load_dependency.side_effect = lambda *args, **kwargs: _load_dependency_side_effect(
-        sources, visited, *args
-    )
-    file_loader.is_file.side_effect = is_file_side_effect
-    file_loader.is_notebook.return_value = False
+    file_loader = _local_loader_with_side_effects(FileLoader, sources, visited)
     site_packages = SitePackages.parse(locate_site_packages())
-    builder = DependencyGraphBuilder(DependencyResolver(whi, site_packages, file_loader, LocalNotebookLoader()))
+    provider = SysPathProvider.from_pathlike_string("")
+    notebook_loader = LocalNotebookLoader(provider)
+    dependency_resolver = DependencyResolver(
+        [
+            NotebookResolver(notebook_loader),
+            SitePackagesResolver(site_packages, file_loader, provider),
+            WhitelistResolver(whi),
+            LocalFileResolver(file_loader),
+        ]
+    )
+    builder = DependencyGraphBuilder(dependency_resolver)
     builder.build_local_file_dependency_graph(Path(paths[0]))
-    assert builder.problems == [
+    assert list(builder.problems) == [
         DependencyProblem(
-            'dependency-check', 'Could not locate import: some_library', Path("root7.py.txt"), 1, 0, 1, 19
+            'import-not-found', 'Could not locate import: some_library', Path("root7.py.txt"), 1, 0, 1, 19
         )
     ]
 
@@ -227,23 +296,29 @@ def test_dependency_graph_builder_raises_problem_with_non_constant_notebook_argu
     sources: dict[str, str] = dict(zip(paths, _load_sources(SourceContainer, *paths)))
     visited: dict[str, bool] = {}
     whi = Whitelist()
-
-    def is_file_side_effect(*args):
-        path = args[0]
-        return str(path) in paths
-
-    file_loader = create_autospec(LocalFileLoader)
-    file_loader.load_dependency.side_effect = lambda *args, **kwargs: _load_dependency_side_effect(
-        sources, visited, *args
-    )
-    file_loader.is_file.side_effect = is_file_side_effect
-    file_loader.is_notebook.return_value = False
+    file_loader = _local_loader_with_side_effects(FileLoader, sources, visited)
     site_packages = SitePackages.parse(locate_site_packages())
-    builder = DependencyGraphBuilder(DependencyResolver(whi, site_packages, file_loader, LocalNotebookLoader()))
+    provider = SysPathProvider.from_pathlike_string("")
+    notebook_loader = LocalNotebookLoader(provider)
+    dependency_resolver = DependencyResolver(
+        [
+            NotebookResolver(notebook_loader),
+            SitePackagesResolver(site_packages, file_loader, provider),
+            WhitelistResolver(whi),
+            LocalFileResolver(file_loader),
+        ]
+    )
+    builder = DependencyGraphBuilder(dependency_resolver)
     builder.build_local_file_dependency_graph(Path(paths[0]))
-    assert builder.problems == [
+    assert list(builder.problems) == [
         DependencyProblem(
-            'dependency-check', "Can't check dependency not provided as a constant", Path(paths[0]), 14, 13, 14, 50
+            'dependency-not-constant',
+            "Can't check dependency not provided as a constant",
+            Path(paths[0]),
+            14,
+            13,
+            14,
+            50,
         )
     ]
 
@@ -253,53 +328,47 @@ def test_dependency_graph_builder_visits_file_dependencies():
     sources: dict[str, str] = dict(zip(paths, _load_sources(SourceContainer, *paths)))
     visited: dict[str, bool] = {}
     whi = whitelist_mock()
-    file_loader = create_autospec(LocalFileLoader)
-    file_loader.load_dependency.side_effect = lambda *args, **kwargs: _load_dependency_side_effect(
-        sources, visited, *args
-    )
-    file_loader.is_file.return_value = True
-    file_loader.is_notebook.return_value = False
+    file_loader = _local_loader_with_side_effects(FileLoader, sources, visited)
     site_packages = SitePackages.parse(locate_site_packages())
-    builder = DependencyGraphBuilder(DependencyResolver(whi, site_packages, file_loader, LocalNotebookLoader()))
+    provider = SysPathProvider.from_pathlike_string("")
+    notebook_loader = LocalNotebookLoader(provider)
+    dependency_resolver = DependencyResolver(
+        [
+            NotebookResolver(notebook_loader),
+            SitePackagesResolver(site_packages, file_loader, provider),
+            WhitelistResolver(whi),
+            LocalFileResolver(file_loader),
+        ]
+    )
+    builder = DependencyGraphBuilder(dependency_resolver)
     builder.build_local_file_dependency_graph(Path("root5.py.txt"))
     assert len(visited) == len(paths)
 
 
-def test_dependency_graph_builder_visits_recursive_file_dependencies():
-    paths = ["root6.py.txt", "root5.py.txt", "leaf4.py.txt"]
-    sources: dict[str, str] = dict(zip(paths, _load_sources(SourceContainer, *paths)))
-    visited: dict[str, bool] = {}
-    whi = whitelist_mock()
-    file_loader = create_autospec(LocalFileLoader)
-    file_loader.load_dependency.side_effect = lambda *args, **kwargs: _load_dependency_side_effect(
-        sources, visited, *args
-    )
-    file_loader.is_file.return_value = True
-    file_loader.is_notebook.return_value = False
-    site_packages = SitePackages.parse(locate_site_packages())
-    builder = DependencyGraphBuilder(DependencyResolver(whi, site_packages, file_loader, LocalNotebookLoader()))
-    builder.build_local_file_dependency_graph(Path("root6.py.txt"))
-    assert len(visited) == len(paths)
-
-
-def test_dependency_graph_builder_ignores_builtin_dependencies():
+def test_dependency_graph_builder_skips_builtin_dependencies():
     paths = ["python_builtins.py.txt"]
     sources: dict[str, str] = dict(zip(paths, _load_sources(SourceContainer, *paths)))
     whi = Whitelist()
-
-    def is_file_side_effect(*args):
-        path = args[0]
-        return str(path) in paths
-
-    file_loader = create_autospec(LocalFileLoader)
-    file_loader.is_file.side_effect = is_file_side_effect
-    file_loader.is_notebook.return_value = False
-    file_loader.load_dependency.side_effect = lambda *args, **kwargs: _load_dependency_side_effect(sources, {}, *args)
+    file_loader = _local_loader_with_side_effects(FileLoader, sources, {})
     site_packages = SitePackages.parse(locate_site_packages())
-    builder = DependencyGraphBuilder(DependencyResolver(whi, site_packages, file_loader, LocalNotebookLoader()))
+    provider = SysPathProvider.from_pathlike_string("")
+    notebook_loader = LocalNotebookLoader(provider)
+    dependency_resolver = DependencyResolver(
+        [
+            NotebookResolver(notebook_loader),
+            SitePackagesResolver(site_packages, file_loader, provider),
+            WhitelistResolver(whi),
+            LocalFileResolver(file_loader),
+        ]
+    )
+    builder = DependencyGraphBuilder(dependency_resolver)
     graph = builder.build_local_file_dependency_graph(Path("python_builtins.py.txt"))
-    assert not graph.locate_dependency(Path("os"))
-    assert not graph.locate_dependency(Path("path"))
+    child = graph.locate_dependency(Path("os"))
+    assert child
+    assert len(child.local_dependencies) == 0
+    child = graph.locate_dependency(Path("pathlib"))
+    assert child
+    assert len(child.local_dependencies) == 0
 
 
 def test_dependency_graph_builder_ignores_known_dependencies():
@@ -307,17 +376,19 @@ def test_dependency_graph_builder_ignores_known_dependencies():
     sources: dict[str, str] = dict(zip(paths, _load_sources(SourceContainer, *paths)))
     datas = _load_sources(SourceContainer, "sample-python-compatibility-catalog.yml")
     whitelist = Whitelist.parse(datas[0])
-
-    def is_file_side_effect(*args):
-        path = args[0]
-        return str(path) in paths
-
-    file_loader = create_autospec(LocalFileLoader)
-    file_loader.is_file.side_effect = is_file_side_effect
-    file_loader.is_notebook.return_value = False
-    file_loader.load_dependency.side_effect = lambda *args, **kwargs: _load_dependency_side_effect(sources, {}, *args)
+    file_loader = _local_loader_with_side_effects(FileLoader, sources, {})
     site_packages = SitePackages.parse(locate_site_packages())
-    builder = DependencyGraphBuilder(DependencyResolver(whitelist, site_packages, file_loader, LocalNotebookLoader()))
+    provider = SysPathProvider.from_pathlike_string("")
+    notebook_loader = LocalNotebookLoader(provider)
+    dependency_resolver = DependencyResolver(
+        [
+            NotebookResolver(notebook_loader),
+            SitePackagesResolver(site_packages, file_loader, provider),
+            WhitelistResolver(whitelist),
+            LocalFileResolver(file_loader),
+        ]
+    )
+    builder = DependencyGraphBuilder(dependency_resolver)
     graph = builder.build_local_file_dependency_graph(Path("python_builtins.py.txt"))
     assert not graph.locate_dependency(Path("databricks"))
 
@@ -327,33 +398,43 @@ def test_dependency_graph_builder_visits_site_packages(empty_index):
     whitelist = Whitelist.parse(datas[0])
     paths = ["import-site-package.py.txt"]
     sources: dict[str, str] = dict(zip(paths, _load_sources(SourceContainer, *paths)))
-    file_loader = create_autospec(LocalFileLoader)
-
-    def is_file_side_effect(*args):
-        path = args[0]
-        filename = path.as_posix()
-        return filename in paths or os.path.isfile(filename)
-
-    file_loader.is_file.side_effect = is_file_side_effect
-    file_loader.is_notebook.return_value = False
-    file_loader.load_dependency.side_effect = lambda *args, **kwargs: _load_dependency_side_effect(sources, {}, *args)
-    site_packages = SitePackages.parse(locate_site_packages())
-    builder = DependencyGraphBuilder(DependencyResolver(whitelist, site_packages, file_loader, LocalNotebookLoader()))
+    provider = SysPathProvider.from_pathlike_string(_samples_path(SourceContainer))
+    file_loader = TestFileLoader(provider, sources)
+    site_packages_path = locate_site_packages()
+    site_packages = SitePackages.parse(site_packages_path)
+    notebook_loader = LocalNotebookLoader(provider)
+    dependency_resolver = DependencyResolver(
+        [
+            NotebookResolver(notebook_loader),
+            SitePackagesResolver(site_packages, file_loader, provider),
+            WhitelistResolver(whitelist),
+            LocalFileResolver(file_loader),
+        ]
+    )
+    builder = DependencyGraphBuilder(dependency_resolver)
     graph = builder.build_local_file_dependency_graph(Path("import-site-package.py.txt"))
-    assert graph.locate_dependency(Path("certifi/core.py"))
+    assert graph.locate_dependency(Path(site_packages_path, "certifi/core.py"))
+    assert not graph.locate_dependency(Path("core.py"))
+    assert not graph.locate_dependency(Path("core"))
 
 
 def test_dependency_graph_builder_raises_problem_with_unfound_root_file(empty_index):
-    sources: dict[str, str] = {}
-    ws = create_autospec(WorkspaceClient)
-    ws.workspace.download.side_effect = lambda *args, **kwargs: _download_side_effect(sources, {}, *args, **kwargs)
-    ws.workspace.get_status.return_value = None
     site_packages = SitePackages.parse(locate_site_packages())
     whi = whitelist_mock()
-    file_loader = create_autospec(LocalFileLoader)
-    builder = DependencyGraphBuilder(DependencyResolver(whi, site_packages, file_loader, LocalNotebookLoader()))
+    file_loader = create_autospec(FileLoader)
+    provider = SysPathProvider.from_pathlike_string("")
+    notebook_loader = LocalNotebookLoader(provider)
+    dependency_resolver = DependencyResolver(
+        [
+            NotebookResolver(notebook_loader),
+            SitePackagesResolver(site_packages, file_loader, provider),
+            WhitelistResolver(whi),
+            LocalFileResolver(file_loader),
+        ]
+    )
+    builder = DependencyGraphBuilder(dependency_resolver)
     builder.build_local_file_dependency_graph(Path("root1.run.py.txt"))
-    assert builder.problems == [DependencyProblem('dependency-check', 'File not found: root1.run.py.txt')]
+    assert list(builder.problems) == [DependencyProblem('file-not-found', 'File not found: root1.run.py.txt')]
     file_loader.load_dependency.assert_not_called()
 
 
@@ -364,10 +445,19 @@ def test_dependency_graph_builder_raises_problem_with_unfound_root_notebook(empt
     ws.workspace.get_status.return_value = None
     site_packages = SitePackages.parse(locate_site_packages())
     whi = whitelist_mock()
-    file_loader = create_autospec(LocalFileLoader)
+    file_loader = create_autospec(FileLoader)
     notebook_loader = create_autospec(LocalNotebookLoader)
     notebook_loader.is_notebook.return_value = False
-    builder = DependencyGraphBuilder(DependencyResolver(whi, site_packages, file_loader, notebook_loader))
+    provider = SysPathProvider.from_pathlike_string("")
+    dependency_resolver = DependencyResolver(
+        [
+            NotebookResolver(notebook_loader),
+            SitePackagesResolver(site_packages, file_loader, provider),
+            WhitelistResolver(whi),
+            LocalFileResolver(file_loader),
+        ]
+    )
+    builder = DependencyGraphBuilder(dependency_resolver)
     builder.build_notebook_dependency_graph(Path("root2.run.py.txt"))
-    assert builder.problems == [DependencyProblem('dependency-check', 'Notebook not found: root2.run.py.txt')]
+    assert list(builder.problems) == [DependencyProblem('notebook-not-found', 'Notebook not found: root2.run.py.txt')]
     file_loader.load_dependency.assert_not_called()
